08-mensajeria-async/04-python-python-topic/publisher/requirements.txt
Flask
flasgger
pika
python-dotenv
gunicorn
